EV-91 fix EventItemViewContainer structure
